added not found exception

+++ back/Library/src/main/java/ru/itmo/library/Coordinates.java
@@ -15,11 +15,11 @@ public class Coordinates {
     /**
      * Координата X.
      */
-    private Long x;
+    private Integer x;
 
     /**
      * Координата Y.
      */
-    private Integer y;
+    private Float y;
 
 }
